matching source values should override constructor parameters default values
